Create project design.
The file project_plan/design_v1.txt includes a rough design of the project written before I have done any coding. I wrote out which classes and modules I'll think I need and their potential methods. There are many things I am unsure of at this point (these are in the design document as questions). As I get into the code the design is likely to update and I'll try to keep the design document up to date as this happens. I wanted a '_v1' of the document so I could look back at how I initially saw the project and compare it to how it ended up. I'm not sure yet if I'll keep multiple version of the design document besides this 1st one and then a 2nd "live" (up to date) version.