Create Board#build_starting_board and associated methods
Board#build_starting_board runs upon instantiation of a Board object. It creates the starting board by first creating a hash with keys representing files (or columns) on the board, their values being empty arrays representing that file's cells. Then it fills the hash with Piece objects or nils to reach a full starting board - I split this into several other methods to keep it concise. Finally it saves this hash to the @current_state instance variable (wish I had a better name for this...). One option that came to mind during this work (which I noted in project_plan/next_steps.txt) is to create a Cell class, then change the board's @current_state to be just an array of 64 elements, each being a Cell. I'll keep this in mind as I move forward incase the hash doesn't work out.