Update Board and methods with new Cell design
I changed the board design to be an array of arrays with each subarray representing a row on the board and each element being a cell. A Cell is an object with @position and @contents. I then updated all the board methods to support this change. I think the design this way will make it easier to write the valid move checks and win checks and other board methods needed for the game.