Create Board#place_piece and Board#place_empty_cell to clean up build board methods
The build board methods prior to this were all pretty messy and there was a lot of duplication. I tried to clean it up by making methods out of the places I found duplicates. Now I have a #place_piece method which takes 4 arguments (is that too many?) Also it technically creates a cell, places a piece into the cell, then adds the cell to the board. Buuuut I'm not really sure how to split that up so its single responsibility. At this stage it seems good enough to move on