Fix knight validation list method
Had to go into the validation methods and flip around the indexes/arguments to make them match and make sense. Before, I had the rook working with (row, column) but the knight with (column, row). I hoped to switch everything to (column, row) because that would be like an x,y plane, but it doesn't work well with the place piece method because that takes columns as letters. I'd have to convert the letters into numbers or something for that to work, which I do not want to do.... So, for now at least, the board array's indexes represent [row][column], or [y][x] if it were an x,y plane.