Complete pawn create valid destinations method
Finished the pawn valid destinations method. Currently it creates all valid destinations from a given position based on color (so up 1 or 2 if white, down if black). It does include diagonals at this point, which I may change in the future. I'm thinking I can have a valid move? type method to check if a diagonal is legal, but I'm not yet sure how to do this.