Add symbols to Piece class and update pretty print
I created a method that adds a symbol an an instance variable to the piece class when it is instantiated. I then updated my pretty print method (in Chess) to test if the symbols worked. They do! Although the 'white' is actually black and vice versa. I may switch this in the future but for now I'm going to leave as is.