Create GameBuilder class and move game setup methods into it from Chess
Before this I had a set of methods in Chess to set up the game (like building the board, building all the pieces, the players, etc.) Chess was getting really long so I instead moved all these into a GameBuilder class. A name of just Builder or even Factory (is this a factory) may be better but for now I'm just going to leave it as GameBuilder.

Now, Chess only saves a new GameBuilder object to @game when its instantiated, then talks to other object through game (eg @game.interface.display_welcome_message or w/e). This seems to be a bit cleaner than it was before.