Rearchitecting object relationships, among other changes....
Most of today I've spent thinking about how best to architect the object relationships in this damn game. What I've done at this point is make a Piece superclass and 8 subclasses for each type. I then wrote out a few different solutions for the move validation problem. The one I've landed on at the end of today is to have each piece type implement its own #available_moves method which takes in a starting_position and board_state argument, then returns a list of available moves. I considered (and still am) telling each Piece which cell it belongs to (@cell) so that this method doesn't even need a starting_cell, but then both each Cell and Piece would know the piece's position, which is redundant. Further, each Piece's @cell would have to be updated every time it is moved.

I also decided it would be best to rewrite how the GameBuilder builds the game. It seems unnecessary to create the pieces and place them in their starting positions in different steps b/c even though in real life a piece could exist outside of a cell, as far as the game is concerned, it doesn't.

Exact next steps are in next_steps.txt