End of day commit... I'm tired ...
Ended up completing the Rook's #available_destination method, which took a whiiiiiile. The algorithm to figure it out is quite complex and I went through a few different versions of its implementation before landing on a good one. Then, writing the tests for it was pretty difficult as well because it needs a board with cell's and pieces to make its decisions. I just need a few more tests to cover some different cases then I'm done with Rook and its tests.

After Rook, I'll move on to every other piece type. Once I get all those done, I have a feeling that I'll be through the hardest part of this project. Finger's crossed!