Rework GameBuilder#build_starting_board
Before this method would run after another method called #fill_piece_collection which would fill an array with all the necessary piece objects for the game, then #build_starting_board would pull pieces out of that array as it created the board. I thought about this and it just seems unnecessary - while in real life pieces can exist outside of a board, for the game the pieces might as well not exist when they aren't on the board. So, I now have it so the pieces are created and placed on the board in the same step.