Clean up player assignment in GameBuilder
Before when GameBuilder created all the starting pieces, it set their @owner values to strings (eg 'Player 1'). Now it assigns the player1/player2 objects instead. This was needed to fix a problem with the Validator methods where it was trying to compare piece.owner to a player object and always failing even when they should have been the same. ie it was comparing player1 the object to 'Player 1'.