Update check and checkmate validation methods
Fixes to check methods: the loops through the board now loop through each row, then each cell. Before it was trying to loop through each row as if it was a cell so it was failing trying to run cell methods against an array.

Update to checkmate: Added functionality to check if any of a player's pieces moves could stop the check. Before I was just checking if any king moves could take it to a safe position. Now it checks to see if any piece's moves can block or capture the attacker to break the check. The full checkmate method is now quite long. I plan to break it up when I refactor.