Complete load and save features!
The game can now be saved and loaded. Woop woop! The game now begins with a welcome message and then a question to the user if asking if they would like to load a game. I have this sequence in my driver script (main.rb) so that depending on the user's response it either sets the chess object to the loaded game or to a new game. I wanted to have my driver script only run one object creation and one 'play_chess' line but I don't think I can do that with my current setup (I'd have to abstract to another class, maybe LoadManager?, but that seems...idk.. unneccessary? Something to consider when refactoring)