Add loop to ensure player in check gets out of check
At the end of a turn a check now runs to make sure the player playing is not in check. This makes sure that a player who was in check at the beginning of the turn gets themselves out of it. There is already another check during the player action validation sequence to make sure a player doesn't put themselves into check. Having both of these is kinda redundant, but they have different objectives (1 being to make sure playing in check gets out of it, the 2nd being to make sure a player doesn't go into check).

Also added some blank lines to clean up UI and added some save files that are helpful for testing.