Feature: Stalemate
The game now has support for stalemates! After a player makes their action, the game checks if their is a stalemate condition. If so, it alerts the players and ends the game. Interestingly, the logic for this is nearly identical to that of checkmate. The only difference is that checkmate first makes sure the game is in check, while stalemate first makes sure the game is not in check.