Refactor all piece classes
This commit includes many changes for improving all my piece type classes. All of the changes in this commit should not have affected program funcitionalitu. I focused on eliminating as many rubocop violations as I could. There are still a few method lengths and one module length (LineMovement) that I wasn't able to eliminate. I disabled these with comments in my code. I'm not sure if its clean to leave this rubocop disable comments but I like to have them there as a marker of ownership of my inability to improve the code further. I may come back in the future and see what I can do to eliminate them.