Refactor Chess class
I went through and tried to refactor Chess as much as I could. They are still many places for improvement. The turn loop works kinda weird in that it sets the player action in one of two ways depending on if the player is in check, but the two ways are a bit redundant, and the 'in check' method is long and ugly. I think I'd have to rethink how this part of the game works to really clean it up. Also, the #player_action_sequence is too long. I'm not sure how to clean it up in its current state because it loops asking the player for their action then validating against a bunch of conditions, displaying a message and resetting the loop if one fails. Because each validation section sends a 'next' command to the loop, I don't know how to split up the method. I think I would need to rethink how this one works as well in order to clean it up.