Refactor tests
I went back and cleaned up the tests I wrote earlier in the project after all the changes I made. Mainly the Validator class had changed so much I removed most of the tests in it. Ideally I would have been writing and updating tests as I was writing and changing the program, but I didn't do that. I'd like to come back to this project after some time to write tests for each important piece.